feat(ui): add Turing machine halted dialog
Implemented a dialog builder which is triggered when the Turing machine reaches a halted state, to display pertinent information such as total steps and space used, input and output. This is visualized in two different forms, depending on whether there is output or not. FXML files for halted state dialog were also added.